requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.0.2
pgvector==0.2.4
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.0.0
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.14
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
